requirements.txt
matplotlib==3.4.3
numpy==1.21.2
opencv_python==4.5.4.58
Pillow==8.4.0
scikit_image==0.18.3
skimage==0.0
streamlit==1.2.0
torch==1.10.0
torchvision==0.11.1
